requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.7.1
pandas==2.2.2
aiofiles==23.2.1
psycopg2-binary==2.9.9
